Add Text Render Api

+++ skyrim-platform/src/tilted/ui/TextToDraw.h
@@ -9,7 +9,7 @@ struct TextToDraw
   float x = 0.f;
   float y = 0.f;
   std::wstring string;
-  std::array<float, 4> color = { 0.f, 0.f, 1.f, 1.f };
+  std::array<double, 4> color = { 0.f, 0.f, 1.f, 1.f };
 };
 
 using TextToDrawCallback = std::function<void(const TextToDraw& textToDraw)>;
